requirements/requirements.txt
packaging
flask == 3.0.3
werkzeug == 3.0.6
flask-restx >= 1.3.0, < 2.0.0
pandas == 2.2.3
python-multipart == 0.0.18
cryptography>=35.0
python-dotenv >= 0.19.0
psycopg[binary]
waitress >= 1.4.4
psutil
sqlalchemy >= 2.0.0, < 3.0.0
psycopg2-binary  # This is required for using sqlalchemy with postgres
alembic >= 1.3.3
redis >=5.0.0, < 6.0.0
walrus==0.9.3
flask-compress >= 1.0.0
appdirs >= 1.0.0
mindsdb-sql-parser ~= 0.7.1
pydantic ~= 2.7.0
mindsdb-evaluator == 0.0.17
duckdb ~= 1.2.1
requests == 2.32.3
dateparser==1.2.0
dill == 0.3.6
numpy
pytz
botocore
boto3 >= 1.34.131
python-dateutil
scikit-learn==1.5.2
hierarchicalforecast~=0.4.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.35
langchain-experimental==0.3.3
langchain-nvidia-ai-endpoints==0.3.3
langchain-openai==0.3.6
langchain-anthropic==0.2.4
langchain-text-splitters==0.3.2
langchain-google-genai>=2.0.0
lark
lxml==5.3.0 # Required for knowledge base webpage embeddings
pgvector==0.3.6 # Required for knowledge bases
prometheus-client==0.20.0
sentry-sdk[flask] == 2.14.0
openai<2.0.0,>=1.58.1
pyaml==23.12.0
mcp==1.6.0
fastapi>=0.110.0, <1.0.0  # Required for LiteLLM server
uvicorn>=0.30.0, <1.0.0   # Required for LiteLLM server

# files reading
pymupdf==1.25.2
filetype
charset-normalizer
openpyxl # used by pandas to read txt and xlsx files
aipdf==0.0.5